+++ src/main/java/au/com/addstar/serversigns/commands/SubCommandSetLoops.java
@@ -18,13 +18,13 @@ public class SubCommandSetLoops extends SubCommand {
             return;
         }
 
-        if (argInt(1, Integer.valueOf(60)).intValue() < 1) {
+        if (argInt(1, 60) < 1) {
             if (verbose) msg(Message.DELAY_GREATER_THAN_ZERO);
             if (verbose) sendUsage();
             return;
         }
 
-        applyMeta(SVSMetaKey.LOOP, new SVSMetaValue(argInt(0)), new SVSMetaValue(argInt(1, Integer.valueOf(60))));
+        applyMeta(SVSMetaKey.LOOP, new SVSMetaValue(argInt(0)), new SVSMetaValue(argInt(1, 60)));
         if (verbose) msg(Message.RIGHT_CLICK_BIND_LOOPS);
     }
 }
